DATACOUCH-596 - Fix threadsafe parameters test.
Move N1qlCouchbaseRepositoryTests to integration. It requires docker to run.
